+++ slack-archive/src/test/java/juja/microservices/slack/archive/integration/ArchiveRepositoryTest.java
@@ -1,4 +1,4 @@
-package juja.microservices.slack.integration;
+package juja.microservices.slack.archive.integration;
 
 import com.lordofthejars.nosqlunit.annotation.UsingDataSet;
 import juja.microservices.slack.archive.model.Channel;
